requirements.txt
torch
torchvision
transformers
pillow
streamlit

